# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/user/create": {
            "post": {
                "description": "创建用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ADMIN User"
                ],
                "summary": "创建用户",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserAdminCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/admin/user/edit": {
            "post": {
                "description": "更新用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ADMIN User"
                ],
                "summary": "更新用户",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserAdminEditRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/admin/user/find/paging": {
            "post": {
                "description": "查询用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ADMIN User"
                ],
                "summary": "查询用户",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserAdminFindPagingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.CommonResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.UserAdminFindPagingResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/user/remove": {
            "post": {
                "description": "删除用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ADMIN User"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserAdminRemoveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.UserAdminCreateRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "description": "用户密码",
                    "type": "string"
                },
                "user_type": {
                    "description": "用户类型 1-admin，2-user",
                    "type": "integer"
                },
                "username": {
                    "description": "用户账户",
                    "type": "string"
                }
            }
        },
        "request.UserAdminEditRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "用户id",
                    "type": "integer"
                },
                "password": {
                    "description": "用户密码",
                    "type": "string"
                },
                "user_type": {
                    "description": "用户类型 1-admin，2-user",
                    "type": "integer"
                },
                "username": {
                    "description": "用户账户",
                    "type": "string"
                }
            }
        },
        "request.UserAdminFindPagingRequest": {
            "type": "object",
            "properties": {
                "page_index": {
                    "description": "分页索引",
                    "type": "integer"
                },
                "page_size": {
                    "description": "分页大小",
                    "type": "integer"
                },
                "user_type": {
                    "description": "用户类型 1-admin，2-user",
                    "type": "integer"
                },
                "username": {
                    "description": "用户账户",
                    "type": "string"
                }
            }
        },
        "request.UserAdminRemoveRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "用户id",
                    "type": "integer"
                }
            }
        },
        "response.CommonResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "response.UserAdminFindPagingResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.UserAdminFindPagingResponseData"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "response.UserAdminFindPagingResponseData": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_type": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}